Rearrange tests a bit
